+++ Documentation/changelog.rst
@@ -11,6 +11,9 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 `Unreleased <https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.3.0...HEAD>`_
 --------------------------------------------------------------------------------------------
 
+[1.4.0] - tbd.
+--------------
+
 Added
 ^^^^^